mock-server支持nodejs-es5 升级依赖："mockjs": "^1.1.0", 删除依赖： "@babel/core": "7.0.0", "@babel/register": "7.0.0", "babel-core": "7.0.0-bridge.0", 可支持依赖：https://github.com/mengbaiai/vue-h5-template/blob/vue-h5-template/package.json
