+++ mock/index.js
@@ -1,18 +1,15 @@
-import Mock from 'mockjs'
-import { param2Obj } from '../src/utils'
+const Mock = require('mockjs')
 
-import user from './user'
-import table from './table'
+const user = require('./user')
+const table = require('./table')
 
-const mocks = [
-  ...user,
-  ...table
-]
+const mocks = [].concat(user.default, table.default)
 
 // for front mock
 // please use it cautiously, it will redefine XMLHttpRequest,
 // which will cause many of your third-party libraries to be invalidated(like progress event).
-export function mockXHR() {
+
+exports.mockXHR = function mockXHR() {
   // mock patch
   // https://github.com/nuysoft/Mock/issues/300
   Mock.XHR.prototype.proxy_send = Mock.XHR.prototype.send
@@ -50,18 +47,20 @@ export function mockXHR() {
   }
 }
 
-// for mock server
-const responseFake = (url, type, respond) => {
-  return {
-    url: new RegExp(`${process.env.VUE_APP_BASE_API}${url}`),
-    type: type || 'get',
-    response(req, res) {
-      console.log('request invoke:' + req.path)
-      res.json(Mock.mock(respond instanceof Function ? respond(req, res) : respond))
-    }
+function param2Obj(url) {
+  const search = url.split('?')[1]
+  if (!search) {
+    return {}
   }
+  return JSON.parse(
+    '{"' +
+      decodeURIComponent(search)
+        .replace(/"/g, '\\"')
+        .replace(/&/g, '","')
+        .replace(/=/g, '":"')
+        .replace(/\+/g, ' ') +
+      '"}'
+  )
 }
 
-export default mocks.map(route => {
-  return responseFake(route.url, route.type, route.response)
-})
+exports.default = mocks
