+++ mock/table.js
@@ -1,4 +1,4 @@
-import Mock from 'mockjs'
+const Mock = require('mockjs')
 
 const data = Mock.mock({
   'items|30': [{
@@ -11,7 +11,7 @@ const data = Mock.mock({
   }]
 })
 
-export default [
+exports.default = [
   {
     url: '/vue-admin-template/table/list',
     type: 'get',
